댓글 삭제 전 확인 알림

+++ src/containers/CommentRepliesContainer.js
@@ -15,7 +15,10 @@ function CommentRepliesContainer({ id }) {
 
   // 댓글 삭제
   const onRemove = id => {
-    dispatch(removeComment(id));
+    // eslint-disable-next-line no-restricted-globals
+    if (confirm("댓글을 삭제하시겠습니까?")) {
+      dispatch(removeComment(id));
+    }
   };
 
   return <CommentsList comments={replyComments} onClickLike={onClickLike} onRemove={onRemove} />;
